Rename and update import entry test
